+++ app/src/main/java/com/elian/computeit/core/util/extensions/TestDataExt.kt
@@ -6,7 +6,7 @@ import com.elian.computeit.core.data.model.TestData
 fun TestData.getListOfValuePerSecond(
 	countCondition: (OperationData) -> Boolean = { true },
 	getValue: (currentSecond: Int, countSinceStart: Int) -> Float,
-): Map<Int, Int> {
+): Map<Int, Float> {
 	return (1..timeInSeconds).associateWith { currentSecond ->
 
 		val countSinceStart = listOfOperationData.count {
@@ -15,6 +15,6 @@ fun TestData.getListOfValuePerSecond(
 
 		val value = getValue(currentSecond, countSinceStart)
 
-		value.ifNaNReturnZero().toInt()
+		value.ifNaNReturnZero()
 	}
 }
